requirements.txt
Flask==3.0.0
Pillow==10.1.0
gunicorn==21.2.0
requests==2.31.0
beautifulsoup4==4.12.2
google-cloud-vision==3.5.0
